amazon_rekognition/requirements.txt
asgiref==3.4.1
boto3==1.20.38
botocore==1.23.38
certifi==2020.6.20
Django==4.0.1
image==1.5.33
jmespath==0.10.0
Pillow==9.0.0
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
yapf==0.32.0
